requirements.txt
altgraph
bgtunnel
click
Envelopes
mysql-tunnel @ git+https://github.com/northriverboats/mysql-tunnel@fb06b6f3949890a0cb44435dc55de4ffcd6cdc71
mysqlclient
pkg_resources
pyinstaller
pyinstaller-hooks-contrib
python-dotenv
zipp
